plutus-certification: Add --local mode.
This enables running the service without Cicero, with jobs run in the
background of the server. This is not safe on a real service, since it
runs user-submitted code in-process!

This also adds applications to dapps-certification-interface, which
can be used locally and will soon be used in
dapps-certification-actions instead of the shell scripts:

generate-flake: Generate the certification flake for a given flakeref,
locking it first
build-flake: Build a flake generated by generate-flake
run-certify: Run the certificaton binary generated by build-flake

So if these are run in order with the correct arguments, you can
simulate the pipeline without having any services running at all.