DAC-253: Add projected work into status messages.
In the "certifying" state, there may be a new "plan" field with a list
of CertificationTask objects corresponding to the tasks that will run
during the job.

The "qc-progress" and "qc-result" fields may have a new "expected"
field with a total number of QuickCheck tests to run in that task.

A CertificationTask object has keys "name" and "index", where "index"
is an integer task offset and "name" is either one of a fixed list of
known strings (e.g. "unit-tests" or "standard-property") or an object
with key "unknown-name" set equal to the default Haskell rendering of
the unknown task name.